Inject github api with hilt
