+++ zeapp/android/src/main/java/de/berlindroid/zeapp/zeui/zeabout/ZeAboutViewModel.kt
@@ -3,7 +3,7 @@ package de.berlindroid.zeapp.zeui.zeabout
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import de.berlindroid.zeapp.zeservices.ZeContributorsService
+import de.berlindroid.zeapp.zeservices.github.ZeContributorsService
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
